Phase 1c+1d: wire the MPS core into both public seams (FA2 five-function ABI + FA4 cute dispatch)
Seam A — FA2 (flash_attn/mps/fa2_backend.py): implement fwd / bwd /
varlen_fwd / varlen_bwd / fwd_kvcache over the shared parity-tested core.
- bwd/varlen_bwd fill the caller's preallocated dq/dk/dv IN PLACE
  (including the packed variants' non-contiguous dkv[:, :, i] views) and
  are recompute + torch.autograd.grad over the memory-bounded core.
- lse sign flip: core emits -inf on fully-masked rows (FA4 convention,
  cute/softmax.py:225); the FA2 CUDA contract is +inf
  (csrc/flash_attn/src/softmax.h:180) — flipped here, and only here.
- FA2 window encoding (negative == infinite, flash_api.cpp:141) translated
  to the core's None-based encoding at the seam.
- kvcache: in-place cache append, cache_batch_idx, cache_leftpad.
- perf knobs (num_splits/deterministic/zero_tensors) accepted and ignored;
  dropout, paged KV (block_table), rotary-in-kvcache, return_attn_probs
  raise NotImplementedError naming the feature — never a wrong number.
  Dropout in particular: the CUDA Philox RNG cannot be bit-matched on MPS
  and an unreproducible mask would silently corrupt the recompute backward.

Seam B — FA4 (flash_attn/cute/interface.py + flash_attn/mps/fa4_backend.py):
flash_attn_func / flash_attn_varlen_func dispatch MPS tensors to the core
ABOVE the autograd.Function, so autograd provides the backward (incl. dlse).
(out, lse) 2-tuple contract preserved; window resolution reuses the
interface's own _resolve_causal_local_window (left+right<0 collapse quirk
included); lse keeps the native -inf convention. cute-only features
(score_mod/mask_mod, qv/MLA, page_table, block sparsity, gather_kv_indices)
raise NotImplementedError. CUDA paths untouched.

Shared varlen driver (flash_attn/mps/varlen.py): differentiable
per-sequence loop around the dense core — packed 3D + cu_seqlens
(+ seqused caps) and batched 4D + seqused. Correct first, fast never mind.

Core precision fix (flash_attn/mps/core.py): the recompute backward now
uses fp32 k/v leaves so per-Q-chunk dK/dV gradients are accumulated in
fp32 and rounded to the storage dtype exactly once. Found by the repo's
own test_flash_attn_splitkv (alibi, seqlen_k=1) on MPS: per-chunk fp16
rounding before accumulation cost 2+ ulp. Pinned by
tests/mps/test_attention_parity.py::test_bwd_chunk_accumulation_exact.

flash_attn_interface.py: on MPS the _wrapped_* aliases call the raw Python
implementations instead of torch.ops.* — torch.library.custom_op runs its
backend impl below the Autograd dispatch key, which silently breaks the
recompute-based backward. Ops are still registered; CUDA binding unchanged.

tests/test_flash_attn.py made device-parametrizable (FLASH_ATTN_TEST_DEVICE,
default "cuda" — CUDA behavior identical): guarded module-scope
get_device_capability calls, guarded triton-only rotary import, 16GB-VRAM
skip helper.

Verification on M-series (all post-fix):
- tests/mps: 888 passed (790 core parity + 98 new seam tests) + 2 new
  regression tests. Seam tests pin the in-place dq/dk/dv fill, both lse
  conventions, the window quirks, kvcache mutation, and the raise matrix.
- repo FA2 suite on MPS: 1592-test dropout-free subset all green
  (1575 run post-fix + race_condition reduced to small seqlens — the
  250-iteration batch-60 loops take hours unfused; documented).
- e2e smoke: flash_attn_func + flash_attn_varlen_func + cute funcs on MPS,
  fwd+bwd, grads finite and matching the CPU fp32 oracle.

docs/apple_silicon/MPS_STATUS.md: the honesty document — per-seam feature
matrix (supported / degraded / unsupported), the deliberate masked-row lse
divergence, and the performance caveats stated plainly.

+++ flash_attn/mps/core.py
@@ -379,8 +379,14 @@ class MPSFlashAttnFunc(torch.autograd.Function):
         q_chunk_size = ctx.q_chunk_size
         if dout is None:  # only lse was used downstream
             dout = q.new_zeros((q.shape[0], q.shape[1], q.shape[2], v.shape[-1]))
-        k_leaf = k.detach().requires_grad_()
-        v_leaf = v.detach().requires_grad_()
+        # fp32 leaves (fp16/bf16 -> fp32 is exact, so the recomputed forward is
+        # bit-identical): autograd.grad then returns per-chunk dK/dV in fp32,
+        # so the cross-chunk accumulation below is exact and dK/dV get rounded
+        # to the storage dtype exactly once, at the end. With low-precision
+        # leaves each chunk's gradient would be rounded to fp16/bf16 *before*
+        # accumulation, leaking chunk-boundary rounding into multi-chunk sums.
+        k_leaf = k.detach().float().requires_grad_()
+        v_leaf = v.detach().float().requires_grad_()
         dq_chunks = []
         dk_acc = torch.zeros(k.shape, dtype=torch.float32, device=k.device)
         dv_acc = torch.zeros(v.shape, dtype=torch.float32, device=v.device)
